Add files via upload
Objective:
Simulate memory allocation with hole-fitting algorithms (First-fit, Best-fit) and implement de-allocation and defragmentation using a linked list.

Specification:
The program simulates memory allocation with a chosen hole-fitting algorithm (First-fit, Best-fit) and implements de-allocation and defragmentation. A menu controls the operations, and each choice calls the appropriate procedure, where the choices are: 1) Enter parameters; 2) Allocate memory for a block; 3) Deallocate memory for a block; 4) Defragment memory; 5) Quit program and free memory.

Assignment:
The size of physical memory is represented by an integer pmsize.
The allocated blocks are contained within a linked list, where each allocated block is a structure containing an id, the starting address, the ending address, and a link to the next allocated block.
Each allocation request specifies the id and the size of the new block. If the id is a duplicate and/or the remaining physical memory is not enough to fit the request, the request is rejected.
Each de-allocation request specifies the id. If the id is invalid, the request is rejected.
Defragmentation compacts the blocks to be contiguous and coalesces the holes into one hole at the far--right end (highest memory addresses) of physical memory.